+++ StuHub/app/helpers/sessions_helper.rb
@@ -3,6 +3,7 @@ module SessionsHelper
   # login
   def log_in(user)
     session[:user_id] = user.id
+    user.touch :last_active_at
   end
 
     #remember user