+++ StuHub/db/migrate/20151109111612_change_column_name_in_group.rb
@@ -1,0 +1,11 @@
+class ChangeColumnNameInGroup < ActiveRecord::Migration
+  def up
+    change_column :groups, :name, :string, null: true
+    change_column :groups, :creator, :string, null: true
+  end
+
+  def down
+    change_column :groups, :name, :string, null: false
+    change_column :groups, :creator, :string, null: false
+  end
+end