+++ StuHub/db/migrate/20151109142144_remove_column_from_group_membership.rb
@@ -1,0 +1,9 @@
+class RemoveColumnFromGroupMembership < ActiveRecord::Migration
+  def up
+    remove_column :group_memberships, :join_date
+  end
+
+  def down
+    add_column :group_memberships, :join_date, :datetime
+  end
+end