+++ StuHub/test/models/group_membership_request_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class GroupMembershipRequestTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end